Split off LargeSceneryEntry
