+++ src/openrct2/object/LargeSceneryObject.h
@@ -9,7 +9,7 @@
 
 #pragma once
 
-#include "../world/Scenery.h"
+#include "LargeSceneryEntry.h"
 #include "SceneryObject.h"
 
 #include <memory>
