+++ src/openrct2/rct2/S6Importer.cpp
@@ -44,6 +44,7 @@
 #include "../management/NewsItem.h"
 #include "../management/Research.h"
 #include "../network/network.h"
+#include "../object/LargeSceneryEntry.h"
 #include "../object/ObjectLimits.h"
 #include "../object/ObjectList.h"
 #include "../object/ObjectManager.h"
